Move solver methods to separate module
